added search movie feature and deleted movie model

+++ db/migrate/20220217121358_add_rating_to_movies.rb
@@ -1,5 +1,5 @@
 class AddRatingToMovies < ActiveRecord::Migration[6.1]
   def change
-    add_column :movies, :rating, :integer
+    add_column :movies, :rating, :float
   end
 end
